# frontend/admin/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "RayFlow API",
    "description": "🐟 RayFlow - Flink/Fluss/Paimon 任务发布管理平台 API 文档",
    "contact": {
      "name": "RayFlow Team"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "0.0.4"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/flink/jobs/{id}": {
      "get": {
        "tags": [
          "Flink Job Management"
        ],
        "summary": "获取作业详情",
        "operationId": "getFlinkJob",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FlinkJobResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Flink Job Management"
        ],
        "summary": "更新作业",
        "operationId": "updateFlinkJob",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlinkJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Flink Job Management"
        ],
        "summary": "删除作业",
        "operationId": "deleteFlinkJob",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/flink/jobs": {
      "get": {
        "tags": [
          "Flink Job Management"
        ],
        "summary": "分页查询作业列表",
        "operationId": "listFlinkJobs",
        "parameters": [
          {
            "name": "is_pager",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "job_name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "job_tags",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "job_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "runtime_mode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseFlinkJobResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Flink Job Management"
        ],
        "summary": "创建作业",
        "operationId": "createFlinkJob",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlinkJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FlinkJobResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/flink/jobs/{id}:triggerSavepoint": {
      "post": {
        "tags": [
          "Flink Job Management"
        ],
        "summary": "触发 Savepoint",
        "operationId": "triggerFlinkSavepoint",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavepointRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FlinkSavepointResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/flink/jobs/{id}:start": {
      "post": {
        "tags": [
          "Flink Job Management"
        ],
        "summary": "启动作业",
        "operationId": "startFlinkJob",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FlinkJobResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/flink/jobs/{id}:publish": {
      "post": {
        "tags": [
          "Flink Job Management"
        ],
        "summary": "发布作业版本",
        "operationId": "publishFlinkJob",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlinkJobVersionRequest"
              }
            }
          }
        },
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FlinkJobVersionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/flink/jobs/{id}:debug": {
      "post": {
        "tags": [
          "Flink Job Management"
        ],
        "summary": "调试作业草稿",
        "operationId": "debugFlinkJob",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FlinkJobResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/flink/jobs/{id}:cancel": {
      "post": {
        "tags": [
          "Flink Job Management"
        ],
        "summary": "取消作业",
        "operationId": "cancelFlinkJob",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/flink/jobs/{id}/versions": {
      "get": {
        "tags": [
          "Flink Job Management"
        ],
        "summary": "查询作业版本列表",
        "operationId": "listFlinkJobVersions",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FlinkJobVersionResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/flink/jobs/{id}/savepoints": {
      "get": {
        "tags": [
          "Flink Job Management"
        ],
        "summary": "查询作业 Savepoint 列表",
        "operationId": "listFlinkSavepoints",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FlinkSavepointRecordResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/flink/jobs/{id}/executions": {
      "get": {
        "tags": [
          "Flink Job Management"
        ],
        "summary": "查询作业执行历史",
        "operationId": "listFlinkJobExecutions",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FlinkJobExecutionResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/flink/jobs/{id}/checkpoints": {
      "get": {
        "tags": [
          "Flink Job Management"
        ],
        "summary": "查询作业 Checkpoint 状态",
        "operationId": "getFlinkCheckpoints",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FlinkCheckpointResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FlinkJobRequest": {
        "required": [
          "jobName"
        ],
        "type": "object",
        "properties": {
          "jobName": {
            "maxLength": 256,
            "minLength": 0,
            "type": "string"
          },
          "jobGroup": {
            "maxLength": 128,
            "minLength": 0,
            "type": "string"
          },
          "jobType": {
            "pattern": "JAR|SQL",
            "type": "string"
          },
          "runtimeMode": {
            "pattern": "BATCH|STREAMING",
            "type": "string"
          },
          "submitType": {
            "pattern": "REST|SQL_GATEWAY|K8S_APPLICATION",
            "type": "string"
          },
          "executionMode": {
            "pattern": "standalone|k8s-application",
            "type": "string"
          },
          "clusterId": {
            "type": "integer",
            "format": "int64"
          },
          "content": {
            "type": "string"
          },
          "mainClass": {
            "type": "string"
          },
          "args": {
            "type": "string"
          },
          "flinkConfig": {
            "type": "string"
          },
          "parallelism": {
            "maximum": 1024,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "type": "string"
          },
          "jobTags": {
            "maxLength": 512,
            "minLength": 0,
            "type": "string"
          },
          "docUrl": {
            "maxLength": 1024,
            "minLength": 0,
            "type": "string"
          },
          "savepointPath": {
            "type": "string"
          },
          "applicationImage": {
            "maxLength": 256,
            "minLength": 0,
            "type": "string"
          },
          "jarUri": {
            "maxLength": 512,
            "minLength": 0,
            "type": "string"
          },
          "dependencyRefs": {
            "type": "string"
          },
          "alertChannelId": {
            "type": "integer",
            "format": "int64"
          },
          "alertRule": {
            "type": "string"
          }
        }
      },
...
      "FlinkJobResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "jobName": {
            "type": "string"
          },
          "jobGroup": {
            "type": "string"
          },
          "jobType": {
            "type": "string"
          },
          "runtimeMode": {
            "type": "string"
          },
          "submitType": {
            "type": "string"
          },
          "executionMode": {
            "type": "string"
          },
          "clusterId": {
            "type": "integer",
            "format": "int64"
          },
          "currentExecutionId": {
            "type": "integer",
            "format": "int64"
          },
          "flinkJobId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "publishStatus": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "mainClass": {
            "type": "string"
          },
          "args": {
            "type": "string"
          },
          "flinkConfig": {
            "type": "string"
          },
          "parallelism": {
            "type": "integer",
            "format": "int32"
          },
          "savepointPath": {
            "type": "string"
          },
          "applicationImage": {
            "type": "string"
          },
          "jarUri": {
            "type": "string"
          },
          "dependencyRefs": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "jobTags": {
            "type": "string"
          },
          "docUrl": {
            "type": "string"
          },
          "alertChannelId": {
            "type": "integer",
            "format": "int64"
          },
          "alertRule": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SavepointRequest": {
        "type": "object",
        "properties": {
          "targetDirectory": {
            "type": "string"
          },
          "cancelJob": {
            "type": "boolean"
          }
        }
      },
      "FlinkSavepointResponse": {
        "type": "object",
        "properties": {
          "recordId": {
            "type": "integer",
            "format": "int64"
          },
          "requestId": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "flinkJobId": {
            "type": "string"
          },
          "targetDirectory": {
            "type": "string"
          },
          "cancelJob": {
            "type": "boolean"
          }
        }
      },
      "FlinkJobVersionRequest": {
        "type": "object",
        "properties": {
          "remark": {
            "type": "string"
          }
        }
      },
      "FlinkJobVersionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "jobId": {
            "type": "integer",
            "format": "int64"
          },
          "versionNo": {
            "type": "integer",
            "format": "int32"
          },
          "versionName": {
            "type": "string"
          },
          "remark": {
            "type": "string"
          },
          "imageUri": {
            "type": "string"
          },
          "imageDigest": {
            "type": "string"
          },
          "imagePublishStatus": {
            "type": "string"
          },
          "imagePublishLog": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PageResponseFlinkJobResponse": {
        "type": "object",
        "properties": {
          "list": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FlinkJobResponse"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        }
      },
      "FlinkSavepointRecordResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "jobId": {
            "type": "integer",
            "format": "int64"
          },
          "jobName": {
            "type": "string"
          },
          "flinkJobId": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "targetDirectory": {
            "type": "string"
          },
          "cancelJob": {
            "type": "boolean"
          },
          "status": {
            "type": "string"
          },
          "triggerMessage": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
...
}